Rotary controller.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@3068 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/controlruler.cpp
@@ -77,7 +77,7 @@ ControlRuler::~ControlRuler()
 void
 ControlRuler::slotScrollHoriz(int x)
 {
-    m_currentXOffset = (double(-x)) / getHScaleFactor();
+    m_currentXOffset = int((double(-x)) / getHScaleFactor());
     repaint();
 }
